Commit before merge

+++ core/src/com/mygdx/gamejam/Settings.java
@@ -60,7 +60,7 @@ public class Settings {
 
 	public static int ICE_TIMING = 5;
 	
-	public static int DELAY_GAMEOVER = 1;
+	public static float DELAY_GAMEOVER = 1f;
 	
 	public static float DAY_LENGTH = 6f;
